Sync with titiler (v0.7.0), WIP
WIP: When running this. There is an issue that the CustomPathParams with two members
     sceneid and scene_metadata is passed as path_dependency.
     The user of that dependency assumes to get a string that is interpreted to be a sceneid, not get an object.
     **I would need some gudience on what changes to make.**

Update version of (this) package to 0.1.0
Use rio-tiler-pds and titiler 0.7.0.
Advance aws-cdk version to 1.160.0

The titiler package has been split into three packages, do refer to each of them.
Update python imports and usage.

(When trying to build using the old Dockerfile, the output was too big to be used in the old way.)
Change to build a docker imaged based on AWS python image instead of creating a bundle that is
overlayed on the AWS python environment.

Changed several things in stack/app.py:
* One thing had changed name to apigw_integrations.HttpLambdaIntegration
  and required a new id argument.
* Use a docker image instead of using docker image to construct a usable zip for the python runtime
  FROM_IMAGE.